requirements.txt
fastapi
uvicorn
Pillow
numpy
onnxruntime
opencv-python
ultralytics
Streamlit
python-multipart